Successfully tested the crawler on live internet (safe domains) to prove its working. The DB connection , redis coordination layer , multiple workers model, Admin REPL, page downloading, page parsing, extraction of links , words , media, periodic batch sync and shutdown and recovery logic are working fine. Also added a periodic snapshot or dumping mechanism for fault tolerance in case of power or DB outages and some minor fixes.

+++ web_crawler/pipeline/synchronizer/sync_crawlqueue.py
@@ -1,6 +1,6 @@
 import redis
-from crawler_exceptions.CrawlerDBErr import MysqlPoolErr
-from crawler_exceptions.CrawlerDBErr import RedisPoolErr
+from web_crawler.crawler_exceptions.CrawlerDBErr import MysqlPoolErr
+from web_crawler.crawler_exceptions.CrawlerDBErr import RedisPoolErr
 
 def sync_crawlqueue(workerstate):
     redis_client = workerstate.redis_client
@@ -10,17 +10,18 @@ def sync_crawlqueue(workerstate):
         formatstring = ",".join(["%s"]*len(successful_urls))
 
         #delete succesfully crawled urls from crawl_queue
-        mysql_cursor.execute(f"""DELETE FROM crawl_queue WHERE url IN ({formatstring})""",
+        if successful_urls:
+            mysql_cursor.execute(f"""DELETE FROM crawl_queue WHERE url IN ({formatstring})""",
                              successful_urls)
         
         #update new urls to crawl_queue
 
-        crawling_urls = [url for url in redis_client.lrange("crawl_queue",0,-1)]
+        crawling_urls = [(url,) for url in redis_client.lrange("crawl_queue",0,-1)]
 
-        mysql_cursor.executemany("""INSERT INTO crawl_queue(url,status,in_process_started)"
+        mysql_cursor.executemany("""INSERT INTO crawl_queue(url,status,in_process_started)
                                  VALUES (%s,'in_process',NOW())
                                  ON DUPLICATE KEY UPDATE 
-                                 status ='in_process' 
+                                 status ='in_process',
                                  in_process_started = NOW()""",crawling_urls)
     
     except redis.RedisError as err:
